requirements.txt
--find-links https://download.pytorch.org/whl/torch_stable.html
fire==0.3.1
numpy==1.18.5
torch==1.7.1+cu110
torchnet==0.0.4
tqdm==4.48.2
visdom==0.1.8.9
pyyaml==5.3.1
